wip: split subcommand (init, serve)
