+++ cli/init.ts
@@ -1,0 +1,6 @@
+/**
+ *
+ */
+export function init() {
+  console.debug("init...");
+}
