+++ cli/serve.ts
@@ -1,0 +1,3 @@
+export async function serve(options: unknown) {
+  console.debug("serve...", options);
+}
